meld can be anywhere....
on my machine (mac OS X), my package installer put meld in /usr/local/bin.

I preserved the existing hard-coded checking for meld, and added in looking to see if meld is in the path anywhere. Line 727 says that if the result of 'which meld' is indeed a file that exists, do nothing (and therefore use that path).